+++ src/malloc.rs
@@ -60,12 +60,12 @@ unsafe fn opj_aligned_alloc_n(
 
   /* alignment shall be at least sizeof(void*) */
   assert!(
-    alignment != 0 as libc::c_uint as libc::c_ulong
-      && alignment & alignment.wrapping_sub(1 as libc::c_uint as libc::c_ulong)
-        == 0 as libc::c_uint as libc::c_ulong
+    alignment != 0u64
+      && alignment & alignment.wrapping_sub(1u64)
+        == 0u64
   );
   assert!(alignment >= ::std::mem::size_of::<*mut libc::c_void>() as libc::c_ulong);
-  if size == 0 as libc::c_uint as libc::c_ulong {
+  if size == 0u64 {
     /* prevent implementation defined behavior of realloc */
     return 0 as *mut libc::c_void;
   }
@@ -90,12 +90,12 @@ unsafe fn opj_aligned_realloc_n(
 
   /* alignment shall be at least sizeof(void*) */
   assert!(
-    alignment != 0 as libc::c_uint as libc::c_ulong
-      && alignment & alignment.wrapping_sub(1 as libc::c_uint as libc::c_ulong)
-        == 0 as libc::c_uint as libc::c_ulong
+    alignment != 0u64
+      && alignment & alignment.wrapping_sub(1u64)
+        == 0u64
   );
   assert!(alignment >= ::std::mem::size_of::<*mut libc::c_void>() as libc::c_ulong);
-  if new_size == 0 as libc::c_uint as libc::c_ulong {
+  if new_size == 0u64 {
     /* prevent implementation defined behavior of realloc */
     return 0 as *mut libc::c_void;
   }
@@ -104,8 +104,8 @@ unsafe fn opj_aligned_realloc_n(
   r_ptr = realloc(ptr, new_size); /* fast path */
   /* we simply use `size_t` to cast, since we are only interest in binary AND
    * operator */
-  if r_ptr as size_t & alignment.wrapping_sub(1 as libc::c_uint as libc::c_ulong)
-    != 0 as libc::c_uint as libc::c_ulong
+  if r_ptr as size_t & alignment.wrapping_sub(1u64)
+    != 0u64
   {
     /* this is non-trivial to implement a portable aligned realloc, so use a
      * simple approach where we do not need a function that return the size of an
@@ -123,7 +123,7 @@ unsafe fn opj_aligned_realloc_n(
 }
 #[no_mangle]
 pub(crate) unsafe fn opj_malloc(mut size: size_t) -> *mut libc::c_void {
-  if size == 0 as libc::c_uint as libc::c_ulong {
+  if size == 0u64 {
     /* prevent implementation defined behavior of realloc */
     return 0 as *mut libc::c_void;
   }
@@ -131,7 +131,7 @@ pub(crate) unsafe fn opj_malloc(mut size: size_t) -> *mut libc::c_void {
 }
 #[no_mangle]
 pub(crate) unsafe fn opj_calloc(mut num: size_t, mut size: size_t) -> *mut libc::c_void {
-  if num == 0 as libc::c_int as libc::c_ulong || size == 0 as libc::c_int as libc::c_ulong {
+  if num == 0u64 || size == 0u64 {
     /* prevent implementation defined behavior of realloc */
     return 0 as *mut libc::c_void;
   }
@@ -139,25 +139,25 @@ pub(crate) unsafe fn opj_calloc(mut num: size_t, mut size: size_t) -> *mut libc:
 }
 #[no_mangle]
 pub(crate) unsafe fn opj_aligned_malloc(mut size: size_t) -> *mut libc::c_void {
-  return opj_aligned_alloc_n(16 as libc::c_uint as size_t, size);
+  return opj_aligned_alloc_n(16u32 as size_t, size);
 }
 #[no_mangle]
 pub(crate) unsafe fn opj_aligned_realloc(
   mut ptr: *mut libc::c_void,
   mut size: size_t,
 ) -> *mut libc::c_void {
-  return opj_aligned_realloc_n(ptr, 16 as libc::c_uint as size_t, size);
+  return opj_aligned_realloc_n(ptr, 16u32 as size_t, size);
 }
 #[no_mangle]
 pub(crate) unsafe fn opj_aligned_32_malloc(mut size: size_t) -> *mut libc::c_void {
-  return opj_aligned_alloc_n(32 as libc::c_uint as size_t, size);
+  return opj_aligned_alloc_n(32u32 as size_t, size);
 }
 #[no_mangle]
 pub(crate) unsafe fn opj_aligned_32_realloc(
   mut ptr: *mut libc::c_void,
   mut size: size_t,
 ) -> *mut libc::c_void {
-  return opj_aligned_realloc_n(ptr, 32 as libc::c_uint as size_t, size);
+  return opj_aligned_realloc_n(ptr, 32u32 as size_t, size);
 }
 #[no_mangle]
 pub(crate) unsafe fn opj_aligned_free(mut ptr: *mut libc::c_void) {
@@ -168,7 +168,7 @@ pub(crate) unsafe fn opj_realloc(
   mut ptr: *mut libc::c_void,
   mut new_size: size_t,
 ) -> *mut libc::c_void {
-  if new_size == 0 as libc::c_uint as libc::c_ulong {
+  if new_size == 0u64 {
     /* prevent implementation defined behavior of realloc */
     return 0 as *mut libc::c_void;
   }
